is_all_npm_packages_installed goes from system to debug
