+++ anvio/utils/system.py
@@ -1,21 +1,11 @@
 import os
-import shutil
+import psutil
 import linecache
 import tracemalloc
 
-import anvio
-
 from anvio.errors import ConfigError
-from anvio.terminal import Run, SuppressAllOutput
 
 from anvio.utils.algorithms import human_readable_file_size
-
-with SuppressAllOutput():
-    try:
-        import psutil
-        PSUTIL_OK=True
-    except:
-        PSUTIL_OK=False
 
 
 def get_total_memory_usage(keep_raw=False):
@@ -27,9 +17,6 @@ def get_total_memory_usage(keep_raw=False):
         A human readable format is returned, e.g. "1.41 GB". If keep_raw, the raw number is
         returned, e.g. 1515601920
     """
-    if not PSUTIL_OK:
-        return None
-
     current_process = psutil.Process(os.getpid())
     mem = current_process.memory_info().rss
     for child in current_process.children(recursive=True):
@@ -39,7 +26,6 @@ def get_total_memory_usage(keep_raw=False):
             pass
 
     return mem if keep_raw else human_readable_file_size(mem)
-
 
 
 def display_top_memory_usage(snapshot, key_type='lineno', limit=10):
@@ -205,50 +191,3 @@ def check_h5py_module():
                           "The reason why the standard anvi'o package does not include this module is both "
                           "complicated and really unimportant. Re-running the migration after `h5py` is installed "
                           "will make things go smoothly.")
-
-
-
-def is_all_npm_packages_installed(run=Run()):
-    """A function to test whether all npm packages are installed in the interactive directory.
-
-    This check is for ensuring that necessary npm packages are installed in the
-    anvio/data/interactive directory.
-    """
-
-    # find the root directory of anvi'o module
-    anvio_module_path = os.path.dirname(os.path.abspath(anvio.__file__))
-    interactive_dir_path = os.path.join(anvio_module_path, 'data', 'interactive')
-
-    if not os.path.exists(interactive_dir_path):
-        raise ConfigError("The interactive directory does not exist in the anvi'o module. "
-                          "Please ensure the directory is present.")
-
-    # Check if Node.js is installed
-    if shutil.which("node") is None:
-        run.warning("It seems your installation is missing Node.js, a recent requirement of anvi'o "
-                    "environments. Please run the following command in your terminal, and you should "
-                    "be good to go:", header="⚠️  YOUR ATTENTION PLEASE ⚠️", overwrite_verbose=True, lc='yellow')
-        run.info_single("      conda install -c conda-forge nodejs", level=0, overwrite_verbose=True, nl_before=1)
-        raise ConfigError("Node.js is not installed. Please install it using conda and try again.")
-
-    # Check if node_modules exists and is not empty
-    node_modules_path = os.path.join(interactive_dir_path, 'node_modules')
-
-    if not os.path.exists(node_modules_path) or not os.listdir(node_modules_path):
-        run.warning("Anvi'o recently changed its use of external libraries for interactive interfaces"
-                    "from git submodules to npm packages. Your current setup does not seem to have the "
-                    "necessary files in place, so the purpose of this warning is to help you match your "
-                    "setup to most up-to-date anvi'o code. If you run the commands below in your terminal, "
-                    "you will most likely be fine :) But if things don't work out, please reach out to us "
-                    "on GitHub or Discord since this is a new feature and some hiccups may occur.",
-                    header="⚠️  YOUR ATTENTION PLEASE ⚠️", overwrite_verbose=True,
-                    lc='yellow')
-        run.info_single(f"1) cd {interactive_dir_path}", level=0, overwrite_verbose=True)
-        run.info_single("2) npm install", level=0, overwrite_verbose=True)
-        run.info_single("3) cd -", level=0, overwrite_verbose=True)
-
-        raise ConfigError("Some npm packages seem to be missing in your interactive directory. "
-                          "Please run 'npm install' in the interactive directory and try again.")
-    else:
-        return True
-
